Fix type mismatches in perf tests to suppress truncation warnings
Change typesize parameter from int to size_t in all getBw function
signatures and the testColl function pointer to match actual usage.
Widen profiler event type fields from int to uint64_t to match the
underlying data. Use size_t for strlen result in ft_test.cu, add
explicit static_cast in getIteration, change getMemString to accept
int64_t, and use jsonSize_t for stepFactor output.

+++ test/perf/all_gather.cu
@@ -67,7 +67,7 @@ testResult_t AllGatherInitData(struct threadArgs* args, ncclDataType_t type, ncc
   return testSuccess;
 }
 
-void AllGatherGetBw(size_t count, int typesize, double sec, double* algBw, double* busBw, int nranks) {
+void AllGatherGetBw(size_t count, size_t typesize, double sec, double* algBw, double* busBw, int nranks) {
   double baseBw = (double)(count * typesize * nranks) / 1.0E9 / sec;
 
   *algBw = baseBw;
